create promise page 컴포넌트 폴더 생성 (컴포넌트 리팩토링)
