+++ src/components/promise/index.js
@@ -1,0 +1,4 @@
+import FindFriend from "./FindFriend";
+import PromiseList from "./PromiseList";
+
+export {PromiseList, FindFriend}
